[#279] Split `Codec.BiMap`
Resolves #279
