+++ test/Test/Toml/Gen.hs
@@ -64,8 +64,7 @@ import GHC.Exts (fromList)
 import Hedgehog (Gen, Range)
 import Numeric.Natural (Natural)
 
-import Toml.Codec.BiMap (toMArray)
-import Toml.Type.AnyValue (AnyValue (..))
+import Toml.Type.AnyValue (AnyValue (..), toMArray)
 import Toml.Type.Key (pattern (:||), Key (..), Piece (..))
 import Toml.Type.PrefixTree (PrefixMap, PrefixTree (..))
 import Toml.Type.TOML (TOML (..))
