+++ tomland.cabal
@@ -81,6 +81,7 @@ library
   exposed-modules:     Toml
                          Toml.Codec
                            Toml.Codec.BiMap
+                             Toml.Codec.BiMap.Conversion
                            Toml.Codec.Code
                            Toml.Codec.Combinators
                            Toml.Codec.Di
@@ -149,6 +150,7 @@ test-suite tomland-test
 
   other-modules:       Test.Toml.Codec
                          Test.Toml.Codec.BiMap
+                           Test.Toml.Codec.BiMap.Conversion
                          Test.Toml.Codec.Combinators
                          Test.Toml.Codec.Generic
